requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
docker==4.4.1
idna==2.10
Markdown==3.3.3
pytz==2021.1
pywin32==227
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
websocket-client==0.57.0
